Add files via upload
# Metal Surface Defect Detection

Metal Surface Defect Detection is a PyQt-based application for detecting defects on metal surfaces using a pre-trained deep learning model.

### Prerequisites

To run the application, you need to have the following dependencies installed:

- Python 3
- PyQt5
- Keras
- PIL (Python Imaging Library)

Usage :
Run the application:
bash
Copy code
python main.py
The application will open with a GUI allowing you to upload an image containing a metal surface with possible defects.

Click on the "UPLOAD IMAGE" button, browse and select the image from your file system.

The application will process the image using the pre-trained model and display the result in the "RESULT" section.

Model
The deep learning model used in this application has been trained on a dataset of metal surface images with various defects. It uses Convolutional Neural Networks (CNN) for defect classification.

Contributing
We welcome contributions to improve this project! If you find any issues or want to add new features, feel free to open an issue or submit a pull request.

License
This project is licensed under the MIT License.